requirements.txt
flask
python-dotenv
requests
google-auth
flask-sqlalchemy 